separate ttl clock
