+++ src/files.cmake
@@ -15,6 +15,7 @@ template.cpp
 ../include/meta/arguments/argument.hpp
 ../include/meta/arguments/packaged_arguments.hpp
 ../include/meta/cache/lru_cache.hpp
+../include/meta/cache/ttl_clock.hpp
 ../include/meta/containers/guarded_sequence_container.hpp
 ../include/meta/containers/iterator.hpp
 ../include/meta/containers/view.hpp
